Made a lot of changes to the connection sequence. You can now use the file picker to choose a .csv to write into. You can't actually write into it yet, but you can create it! You won't believe how aggravating that feature was to implement. Granted, after all the setbacks I've encountered, a couple hours is nothing. The streaming is somewhat more reliable as well.
At this point, I think my goal is to solve the gesture recognition. If I don't get to the game, that's okay, its only real purpose at this point would be as a skin for the gesture recognition. I fear I may run over the deadlines though, as I am having more trouble with Azure and the .csv files than I anticipated.